More refactoring and code improvements
