+++ src/Services/ClearToolbar.ts
@@ -1,0 +1,8 @@
+import { ToolbarState } from "../Toolbar/Toolbar";
+import { AccountData } from "./AccountData";
+
+const ClearToolbar = (state: ToolbarState) => {
+    state.account.Set(new AccountData("", ""))
+}
+
+export { ClearToolbar }
